src/main/java/com/example/FenrisBookShopApp/services/security/BookStoreUserDetailsService.java
package com.example.FenrisBookShopApp.services.security;

import com.example.FenrisBookShopApp.dto.security.BookStoreUserDetails;
import com.example.FenrisBookShopApp.entities.user.UserEntity;
import com.example.FenrisBookShopApp.repositories.UserRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class BookStoreUserDetailsService implements UserDetailsService {
    private final UserRepository userRepository;

    @Override
    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
        UserEntity user = userRepository.findUserByEmail(email);
        if (user == null) {
            throw new UsernameNotFoundException("user not found");
        }

        return new BookStoreUserDetails(user);
    }
}
